+++ .vscode/settings.json
@@ -6,9 +6,11 @@
 		"Alanisposting",
 		"APOD",
 		"BAJALANIS",
+		"bowo",
 		"braincell",
 		"braincells",
 		"burgerboss",
+		"CHABI",
 		"chillbros",
 		"codepaths",
 		"consts",
